chore: synced firewall-consumer with latest audited version

+++ packages/firewall-consumer/contracts/interfaces/IFirewall.sol
@@ -4,13 +4,13 @@
 pragma solidity ^0.8;
 
 interface IFirewall {
-    function preExecution(address sender, bytes memory data, uint value) external;
-    function postExecution(address sender, bytes memory data, uint value) external;
-    function preExecutionPrivateInvariants(address sender, bytes memory data, uint value) external returns (bytes32[] calldata);
+    function preExecution(address sender, bytes memory data, uint256 value) external;
+    function postExecution(address sender, bytes memory data, uint256 value) external;
+    function preExecutionPrivateInvariants(address sender, bytes memory data, uint256 value) external returns (bytes32[] calldata);
     function postExecutionPrivateInvariants(
         address sender,
         bytes memory data,
-        uint value,
+        uint256 value,
         bytes32[] calldata preValues,
         bytes32[] calldata postValues
     ) external;
